requirements.txt
asn1crypto==0.24.0
autopep8==1.4.3
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
cryptography==2.6.1
Django==2.2.1
django-cors-headers==3.0.1
django-rest-knox==4.0.1
djangorestframework==3.9.4
docopt==0.6.2
idna==2.8
psycopg2==2.7.7
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
